move logger initializers to loggers.go
