+++ grohl.go
@@ -1,13 +1,6 @@
 package grohl
 
-import (
-	"io"
-	"os"
-	"os/signal"
-	"path/filepath"
-	"syscall"
-	"time"
-)
+import "time"
 
 type Data map[string]interface{}
 
@@ -48,54 +41,6 @@ func SetLogger(logger Logger) Logger {
 	CurrentContext.Logger = logger
 
 	return logger
-}
-
-func NewChannelLogger(channel chan Data) (*ChannelLogger, chan Data) {
-	if channel == nil {
-		channel = make(chan Data)
-	}
-	return &ChannelLogger{channel}, channel
-}
-
-func NewIoLogger(stream io.Writer) *IoLogger {
-	if stream == nil {
-		stream = os.Stdout
-	}
-
-	return &IoLogger{stream, true}
-}
-
-func NewBufferedLogger(filename string, upperbound int) *BufferedLogger {
-	isFile := len(filename) > 0
-
-	if isFile {
-		os.MkdirAll(filepath.Dir(filename), 0744)
-	}
-
-	return &BufferedLogger{
-		Filename:   filename,
-		Position:   0,
-		Upperbound: upperbound,
-		AddTime:    true,
-		Fallback:   &nopcloser{os.Stdout},
-		buffer:     make([]byte, upperbound),
-		isFile:     isFile,
-	}
-}
-
-func WriteBufferedLogs(logch chan Data, filename string, upperbound int) {
-	logger := NewBufferedLogger(filename, upperbound)
-	sigch := make(chan os.Signal)
-	signal.Notify(sigch, syscall.SIGUSR1, syscall.SIGUSR2, syscall.SIGTERM, syscall.SIGQUIT)
-
-	for {
-		select {
-		case <-sigch:
-			logger.Flush()
-		case data := <-logch:
-			logger.Log(data)
-		}
-	}
 }
 
 func NewContext(data Data) *Context {
